requirements.txt
Django==5.1.3
djangorestframework==3.15.2
PyMySQL==1.1.1
python-decouple==3.8
Requests==2.32.3
